Fix C4267 size_t-to-int narrowing warnings for MSVC build
Add explicit casts where size_t values are passed to int-typed parameters:
- gau2grid_collocation.cxx: cast npts/ncomp to unsigned long at gau2grid
  API boundary; keep size_t for pointer arithmetic to avoid 32-bit overflow
  on Windows (where unsigned long is 32-bit)
- reference_local_host_work_driver.cxx: introduce int32_t locals for all
  size_t values passed to BLAS functions (blas::gemm, blas::dot, etc.)
- Integrator headers: static_cast<int32_t> for task sizes and shell counts
  passed to BLAS and submat map functions

+++ src/xc_integrator/integrator_util/exx_screening.cxx
@@ -128,8 +128,8 @@ void exx_ek_screening(
   // Compute approx F_i^(k) = |P_ij| * B_j^(k) 
   //auto gemm_st = hrt_t::now();
   std::vector<double> task_approx_f( nbf * ntasks );
-  blas::gemm( 'N', 'N', nbf, ntasks, nbf, 1., P_abs, ldp,
-    task_max_bfn.data(), nbf, 0., task_approx_f.data(), nbf );
+  blas::gemm( 'N', 'N', static_cast<int>(nbf), static_cast<int>(ntasks), static_cast<int>(nbf), 1., P_abs, static_cast<int>(ldp),
+    task_max_bfn.data(), static_cast<int>(nbf), 0., task_approx_f.data(), static_cast<int>(nbf) );
   //auto gemm_en = hrt_t::now();
   //std::cout << "... done " << dur_t(gemm_en-gemm_st).count() << std::endl;
